focus on first focusable elm in generic comp
